refactor(agent): the cluster ceremonies await; no block_in_place remains
Last of the three pre-existing block_in_place bridges, and the largest.

`RankTransport` was a sync trait over async peer calls, so `PeerTransport`
ran them with block_in_place + block_on. It now returns boxed futures --
boxed rather than `async fn`, because it is used as `dyn RankTransport`
and async fn in a trait is not dyn-compatible -- and the implementation
awaits each peer call. `ClusterControl` follows for the same reason, since
the driver's own methods had to become async to await the transport.

As with the other two, the runtime `Handle` is DELETED rather than
replaced: with the dials awaited, `PeerTransport` has no use for one, so
its constructor loses that parameter. `atlasctl agent`'s supervisor loop
loses its `spawn_blocking` wrapper too -- it existed precisely because
`supervise()` dialled the network, and it now awaits instead.

`ClusterDriver`'s ceremonies stay ordinary async functions, moved to
`clusterdriver/ceremonies.rs`; the trait methods are a one-line `Box::pin`
each. `session/pairing_address_tests.rs` is split out along the seam it
already had. Both splits are the 500-line cap, not a redesign.

A test caught a real regression this introduced: `supervise` did

    let _ = self.stop_cluster();

which after the change discards a FUTURE. `let _ =` suppresses the
`must_use` warning, so it compiled silently and the survivors were never
stopped -- exactly what `a_rank_that_dies_after_commit_tears_the_cluster_down`
asserts. Now awaited.

Workspace: 817 tests passed, rustfmt clean, clippy zero warnings, no file
over the 500-line cap.

`grep -rn 'tokio::task::block_in_place' crates/` now returns nothing.
