+++ crates/atlasctl-agent/src/clusterdriver/types.rs
@@ -77,7 +77,7 @@ impl ClusterDriver {
     /// Stop the ranks that already started, so a failed commit leaves nothing
     /// running. Failures are ignored for the same reason rollback ignores them:
     /// the operator needs the original error, not this one.
-    pub(crate) fn stop_started(&self, started: &[RankStarted], targets: &[&Target]) {
+    pub(crate) async fn stop_started(&self, started: &[RankStarted], targets: &[&Target]) {
         for r in started {
             let Some(t) = targets.iter().find(|t| t.assignment.node == r.node) else {
                 continue;
@@ -89,7 +89,10 @@ impl ClusterDriver {
                 // Best effort: supervision already knows something is wrong
                 // and has no operator waiting on a return value.
                 Some(addr) => {
-                    let _ = self.transport.stop(t.assignment.node, addr, &r.container);
+                    let _ = self
+                        .transport
+                        .stop(t.assignment.node, addr, &r.container)
+                        .await;
                 }
             }
         }
